Add and group 20 tc wo traffic, add conftest py to vlan dir with setup fixture Signed-off-by: Stepan Vovk <stepan.vovk@plvision.eu>
